requirements.txt
streamlit
yfinance
numpy
pandas
scikit-learn
python-dotenv
openai
